Terms, refactor
